# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API Spec for footbal team management",
        "title": "API Football Team Management",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "afrizalsebastian",
            "email": "sebastiangurning@gmail.com"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "paths": {
...
        "/api/v1/matches": {
            "get": {
                "description": "Get List Matches",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Matches"
                ],
                "summary": "Get List Matches",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Team ID",
                        "name": "teamId",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "type match: one of 'result' or 'scheduled'",
                        "name": "type",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.WebResponse-array_dto_GetMatchResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.WebResponse-any"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.WebResponse-any"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new match",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Matches"
                ],
                "summary": "Create Match",
                "parameters": [
                    {
                        "description": "Create match body request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateMatchRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/api.WebResponse-dto_CreateMatchResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.WebResponse-any"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.WebResponse-any"
                        }
                    }
                }
            }
        },
        "/api/v1/matches/{matchId}": {
            "get": {
                "description": "Get Match Goal list",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Matches"
                ],
                "summary": "Get Match Goal list",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Team ID",
                        "name": "matchId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.WebResponse-dto_MatchGoalsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.WebResponse-any"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.WebResponse-any"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete Match",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Matches"
                ],
                "summary": "Delete Match",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Match ID",
                        "name": "matchId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.WebResponse-any"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.WebResponse-any"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/api.WebResponse-any"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.WebResponse-any"
                        }
                    }
                }
            }
        },
        "/api/v1/matches/{matchId}/full-time": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Submit matches result. Full-Time",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Matches"
                ],
                "summary": "Submit matches result. Full-Time",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Match ID",
                        "name": "matchId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.WebResponse-any"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.WebResponse-any"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/api.WebResponse-any"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.WebResponse-any"
                        }
                    }
                }
            }
        },
        "/api/v1/matches/{matchId}/goals": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new match goal",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Matches"
                ],
                "summary": "Create Match goal",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Match ID",
                        "name": "matchId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Create match body request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.MatchGoal"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/api.WebResponse-any"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.WebResponse-any"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/api.WebResponse-any"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.WebResponse-any"
                        }
                    }
                }
            }
        },
        "/api/v1/matches/{matchId}/reschedule": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Reschedule match",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Matches"
                ],
                "summary": "Reschedule Match",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Match ID",
                        "name": "matchId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Reschedule match request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.RescheduleMatchRequest"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "$ref": "#/definitions/api.WebResponse-any"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.WebResponse-any"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.WebResponse-any"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "api.ErrorsDetail": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "field": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "api.WebResponse-any": {
            "type": "object",
            "properties": {
                "data": {},
                "error_message": {
                    "type": "string"
                },
                "errors_detail": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.ErrorsDetail"
                    }
                },
                "http_code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                },
                "message_code": {
                    "type": "integer"
                },
                "message_id": {
                    "type": "string"
                }
            }
        },
...
        "api.WebResponse-array_dto_GetMatchResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.GetMatchResponse"
                    }
                },
                "error_message": {
                    "type": "string"
                },
                "errors_detail": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.ErrorsDetail"
                    }
                },
                "http_code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                },
                "message_code": {
                    "type": "integer"
                },
                "message_id": {
                    "type": "string"
                }
            }
        },
        "api.WebResponse-dto_CreateMatchResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/dto.CreateMatchResponse"
                },
                "error_message": {
                    "type": "string"
                },
                "errors_detail": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.ErrorsDetail"
                    }
                },
                "http_code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                },
                "message_code": {
                    "type": "integer"
                },
                "message_id": {
                    "type": "string"
                }
            }
        },
...
        "api.WebResponse-dto_MatchGoalsResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/dto.MatchGoalsResponse"
                },
                "error_message": {
                    "type": "string"
                },
                "errors_detail": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.ErrorsDetail"
                    }
                },
                "http_code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                },
                "message_code": {
                    "type": "integer"
                },
                "message_id": {
                    "type": "string"
                }
            }
        },
...
        "dto.CreateMatchRequest": {
            "type": "object",
            "required": [
                "away_team_id",
                "date",
                "home_team_id",
                "time"
            ],
            "properties": {
                "away_team_id": {
                    "type": "string"
                },
                "date": {
                    "type": "string"
                },
                "home_team_id": {
                    "type": "string"
                },
                "time": {
                    "type": "string"
                }
            }
        },
        "dto.CreateMatchResponse": {
            "type": "object",
            "properties": {
                "away_team_id": {
                    "type": "string"
                },
                "date": {
                    "type": "string"
                },
                "home_team_id": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "time": {
                    "type": "string"
                }
            }
        },
...
        "dto.GetListPlayerItem": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "jersey_number": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "position": {
                    "$ref": "#/definitions/dto.PlayerPosition"
                },
                "team": {
                    "$ref": "#/definitions/dto.ListPlayerItemTeam"
                }
            }
        },
...
        "dto.GetMatchResponse": {
            "type": "object",
            "properties": {
                "away_score": {
                    "type": "integer"
                },
                "away_team": {
                    "$ref": "#/definitions/dto.MatchTeam"
                },
                "away_team_id": {
                    "type": "string"
                },
                "date": {
                    "type": "string"
                },
                "goals": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.MatchGoalListItem"
                    }
                },
                "home_score": {
                    "type": "integer"
                },
                "home_team": {
                    "$ref": "#/definitions/dto.MatchTeam"
                },
                "home_team_id": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "time": {
                    "type": "string"
                }
            }
        },
...
        "dto.ListPlayerItemTeam": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "dto.MatchGoal": {
            "type": "object",
            "required": [
                "goal_minute",
                "player_id"
            ],
            "properties": {
                "goal_minute": {
                    "type": "string"
                },
                "player_id": {
                    "type": "string"
                }
            }
        },
        "dto.MatchGoalListItem": {
            "type": "object",
            "properties": {
                "goal_minute": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "player": {
                    "$ref": "#/definitions/dto.GetListPlayerItem"
                }
            }
        },
        "dto.MatchGoalsResponse": {
            "type": "object",
            "properties": {
                "away_team": {
                    "$ref": "#/definitions/dto.ListPlayerItemTeam"
                },
                "goals": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.MatchGoalListItem"
                    }
                },
                "home_team": {
                    "$ref": "#/definitions/dto.ListPlayerItemTeam"
                },
                "match_id": {
                    "type": "string"
                }
            }
        },
        "dto.MatchTeam": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "logo": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "dto.PlayerPosition": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "dto.RescheduleMatchRequest": {
            "type": "object",
            "required": [
                "date",
                "time"
            ],
            "properties": {
                "date": {
                    "type": "string"
                },
                "time": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Please input token with format: Bearer __TOKEN__",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}